Z3dPy v0.3.9
*Nightly build, wiki/examples are based on the release build
Change Notes:
THE RETURN OF OBJECTS

- The larger objects have now been converted to classes, this includes:
Meshes
AniMeshes (which are now just meshes)
Emitters
Cameras
Things

All object functions for these are either redundant and thus removed, or are now functions inside the objects.

Smaller objects like Vectors, Rays, Tris, Dupes, and Particles are still the same for max speed

- Removed WhatIs().


RENDERING

- Raster functions have been renamed to Render functions.
This includes:
Render()
DebugRender()
RenderThings()
RenderMeshList()

The old functions will still be there for backwards compatibility


SHADERS

- Instead of having a separate draw function
for each shade, I'm introducing a shader system.

Each mesh now has a shader property. It's a function
that returns the colour for a given triangle.

For example:

    myMesh = z3dpy.LoadMesh("z3dpy/mesh/suzanne.obj")

    myMesh.shader = SHADER_SIMPLE

    then during the draw loop...

    for tri in z3dpy.RenderMeshList([myMesh]):
        # Now the colour is pre-set.
        pygame.draw.polygon(mySurface, z3dpy.TriGetColour(tri), ...)

        # You may need a VectorFloor() or VectorCeil() to keep things integer-only.


shader has 5 built-in options:

SHADER_UNLIT (default)
    Tri's colour is unchanged

SHADER_SIMPLE
    Basic direction shading.

SHADER_STATIC
    Baked lighting

SHADER_DYNAMIC
    Dynamic lighting

For example:

    def MyShader(tri):
        return z3dpy.TriGetColour(tri)

    myMesh = z3dpy.LoadMesh("mesh/cube.obj")

    myMesh.shader = MyShader


DRAWING

- Completely changed how TriTo__() works, instead of a for loop, the functions take
a draw method.

    def MyDraw(x, y):
        pixelArray[x][y] = z3dpy.TriGetColour(tri)

    while True:
        for tri in z3dpy.Raster():
            z3dpy.TriToPixels(tri, MyDraw)

TriToPixels(tri, draw)
    draw function is given 2 parameters, x, y

TriToLines(tri, draw)
    draw function is given 4 parameters, sx, sy, ex, ey

The current triangle in the draw function can be accessed as 'tri', without having to pass it in.

- Fixed TriToPixels() and TriToLines() visual glitches.

- Added TriToTexels(), which will convert a triangle into pixels with UV coordinates.
    UVs are normalized, so you'll have to multiply and modulo by the size of the texture.

    TriToTexels(tri, draw)
        draw function is given 4 parameters, x, y, u, v


- Removed PgPixelShader() and TkPixelShader(), as there's no longer overhead with
TriToTexels().

- PgDrawTri__() and TkDrawTri__() are now deprecated, though can still be used.

- Added FormatTri(tri, Is1D), which will take a tri and turn it into a list of 2D points for
drawing libraries. Tkinter and Kivy use a 1D array, while PyGame uses a 2D array.


ANIMATED MESHES

- Fixed broken animated meshes

- Created my own Animated OBJ format (.aobj) because the standard method is
inefficient. My method only stores locations of verticies when they change, and
combines it all into one file.
(It may be case-dependent, but in my testing it's a lot smaller than
an OBJ sequence.)

If you don't want to know how the format is laid out, you can skip this next bit

It starts with the first OBJ file as normal, before an 'a' on it's own line, then
it switches to a format of "(vert number) x y z" for each vert that moves, and
a "next" on it's own line when a new frame begins. If the mesh gains verticies
from one frame to the next, there will be a 'new' on it's own line, followed by
the OBJ file of the next frame, before switching back after an 'a'.

OBJ sequences will automatically be converted to an aobj file on loading, with
a printed message about it.

I'm still working out the bugs in my importer, but the format itself has no issues.


MESHES

- Fixed LoadMesh() throwing an error when the file wasn't found, and a "/" was not in the filename.

- Renamed suzanne.obj after realizing my mistake.


Z3DPYFAST

- z3dpyfast is no longer enabled by default. Initialize it with z3dpy.fast()
Though once the functions are replaced, you can't go back to pure Python without resetting.


MISC

- BakedLighting() now requires the list of things. To get the original default behaviour,
put in z3dpy.GatherThings() as the argument.

- Hit functions have been renamed to better match other functions, this includes
HitGetDistance()
HitGetPos()
HitGetNormal()
HitGetWorldPos()

- SetInternalCam() error message displays again, I optimized it too far last time.

- Added ListNDFlatten(list), as looking online it seems like NumPy is the only other option

- Renamed Vec3MatrixMulOneLine() to just Vec3MatrixMul().

- Renamed sign() to Sign(), to fit the rules.

- I finally learned how to stack fors in a single line, prepare yourself.

- Optimizations